pyproject.toml
[tool.poetry]
name = "quizit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.1"
Flask-SocketIO = "4.3.1"
eventlet = "^0.32.0"
PyYAML = "^5.4.1"
python-engineio = "3.13.2"
python-socketio = "4.6.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
pylint = "^2.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
aggressive = 2
